# Change serialization copying with serializers to use expression trees