--- a/Robust.UnitTesting/Shared/Serialization/TypeSerializers/ListSerializerTest.cs
+++ b/Robust.UnitTesting/Shared/Serialization/TypeSerializers/ListSerializerTest.cs
@@ -32,5 +32,30 @@ namespace Robust.UnitTesting.Shared.Serialization.TypeSerializers
 
             Assert.That(deserializedList, Is.EqualTo(list));
         }
+
+        [Test]
+        public void CustomCopyTest()
+        {
+            var source = new List<string> {"A", "E"};
+            var target = new List<string>();
+
+            Assert.IsNotEmpty(source);
+            Assert.IsEmpty(target);
+
+            var copy = (List<string>?) Serialization.CopyWithTypeSerializer(typeof(ListSerializers<string>), source, target);
+
+            Assert.NotNull(copy);
+
+            Assert.IsNotEmpty(copy!);
+            Assert.IsNotEmpty(target);
+
+            Assert.That(copy, Is.EqualTo(target));
+
+            Assert.That(copy, Does.Contain("A"));
+            Assert.That(copy, Does.Contain("E"));
+
+            Assert.That(target, Does.Contain("A"));
+            Assert.That(target, Does.Contain("E"));
+        }
     }
 }