``` ini

BenchmarkDotNet=v0.12.1, OS=Windows 10.0.19042
Intel Core i7-4790K CPU 4.00GHz (Haswell), 1 CPU, 8 logical and 4 physical cores
.NET Core SDK=5.0.201
  [Host]     : .NET Core 5.0.4 (CoreCLR 5.0.421.11614, CoreFX 5.0.421.11614), X64 RyuJIT
  DefaultJob : .NET Core 5.0.4 (CoreCLR 5.0.421.11614, CoreFX 5.0.421.11614), X64 RyuJIT
```


Old
|                      Method |     Mean |     Error |    StdDev |   Median |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|---------------------------- |---------:|----------:|----------:|---------:|-------:|------:|------:|----------:|
| CopyIntegerCustomSerializer | 4.351 μs | 0.1138 μs | 0.3303 μs | 4.241 μs | 0.5569 |     - |     - |   2.28 KB |

---

New
|                      Method |     Mean |   Error |  StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|---------------------------- |---------:|--------:|--------:|-------:|------:|------:|----------:|
| CopyIntegerCustomSerializer | 171.2 ns | 3.36 ns | 3.86 ns | 0.0324 |     - |     - |     136 B |